Create action variant to scan for all vuln types (#7)
Referencing a version tag in this case (rather than a SHA) is considered
okay because it's on the same action. Also, using a specific version tag
rather than the moving major version ref (v1 in this case) because that
would be incorrect when used with older version (e.g. when latest is
v1.0.4 and the user is trying to use v1.0.3).